requirements.txt
yt-dlp==2024.4.9
openai-whisper==20231117
ffmpeg-python==0.2.0
torch==2.2.2